pyproject.toml
[tool.poetry]
name = "footie-fixtures"
version = "0.0.1"
description = "Footie Fixtures"
authors = ["Dom Batten <dominic.batten@googlemail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/dbatten5/footie-fixtures"
repository = "https://github.com/dbatten5/footie-fixtures"
documentation = "https://footie-fixtures.readthedocs.io"
classifiers = [
    "Development Status :: 4 - Beta",
]

[tool.poetry.urls]
Changelog = "https://github.com/dbatten5/footie-fixtures/releases"

[tool.poetry.dependencies]
python = "^3.8.1"
click = ">=8.0.1"
google-api-python-client = "^2.70.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.8.0"

[tool.poetry.group.test.dependencies]
pytest = ">=6.2.5"
coverage = {extras = ["toml"], version = ">=6.2"}
vcrpy = "^4.2.1"

[tool.poetry.group.dev.dependencies]
Pygments = ">=2.10.0"
darglint = ">=1.8.1"
isort = ">=5.10.1"
mypy = "*"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pyupgrade = ">=2.29.1"
safety = "*"
typeguard = ">=2.13.3"
myst-parser = {version = ">=0.16.1"}
black = "*"
types-requests = "*"
ruff = "*"

[tool.poetry.group.docs.dependencies]
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
furo = ">=2021.11.12"

[tool.poetry.scripts]
footie-fixtures = "footie_fixtures.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["footie_fixtures", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = [
  "google.*",
  "google_auth_oauthlib.*",
  "googleapiclient.*",
  "vcr.*"
]
ignore_missing_imports = true

[tool.ruff]
ignore = [
  'B019',
  'D203',
  'D204',
  'D213',
  'D215',
  'D400',
  'D404',
  'D406',
  'D407',
  'D408',
  'D409',
  'D413',
  'E501'
]
line-length = 80
select = [
    'B',
    'B9',
    'C',
    'D',
    'E',
    'F',
    'N',
    'S',
    'W',
]

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.pydocstyle]
convention = 'google'

[tool.ruff.per-file-ignores]
"__init__.py" = ['F401']
"tests/*" = [
    'S101',
    'D212',
    'D415',
    'D205',
    'D104',
    'D100',
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
